+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## 3.44.0
+
+- 6,203 icons (5,147 outline + 1,056 filled)
+- Added 18 new icons
+
+### Tabler Icons release notes
+
+### 18 new icons:
+
+- `outline/code-ai`
+- `outline/email-stamp`
+- `outline/foodsteps`
+- `outline/git-pull-request-conflict`
+- `outline/noise-reduction`
+- `outline/photo-alt`
+- `outline/pointer-2`
+- `outline/pointer-collaboration-2`
+- `outline/pointer-collaboration`
+- `outline/roulette`
+- `outline/scan-cube`
+- `outline/sketching`
+- `outline/sparkle-2`
+- `outline/sparkle-highlight`
+- `outline/sparkle`
+- `outline/sphere-2`
+- `outline/text-scan-ai`
+- `outline/vignette`
+
+Fixed icons: `outline/air-balloon`, `outline/body-scan`, `outline/chart-sankey`, `outline/ear-scan`, `outline/grid-scan`, `outline/line-scan`, `outline/object-scan`, `outline/photo-scan`, `outline/route-scan`, `outline/scan-eye`, `outline/scan-letter-a`, `outline/scan-letter-t`, `outline/scan-position`, `outline/scan-traces`, `outline/scan`, `outline/text-scan-2`, `outline/user-scan`, `outline/zoom-scan`
+
 ## 3.43.0
 
 - 6,185 icons (5,129 outline + 1,056 filled)